Fix compilation errors on removed or incompatible method calls
* Annotate with @CompileStatic

An "Access to log.(debug|info|trace|warn) is forbidden" error occurs on compilation. The workaround is to override getLog() instead of using the @Slf4j annotation. 

* Hack to work around weird User trait inheritance issue

Override with exact copy of `DataModelImporterProviderService.importDomain()` because `IntegrationUser` is not recognised as inheriting from the `User` trait.

Casting `currentUser as User` within `DataModelImporterProviderService` does not work for some reason, hence the code duplication.